♻️ [refactor] #205 : ChallengeItem 컴포넌트 분리
